refactor(api): move issue-types to modules/ as controller, model, service
